Add support for Personal Access Tokens
Harvest have recently added this an alternative form of authentication (see http://help.getharvest.com/api-v2/authentication-api/authentication/authentication/#personal-access-tokens).  This is useful because once generated the token does not expire, meaning the initial call to log the user in is no longer required.  This also means that the user no longer needs to share their username and password.